split table selection state
