+++ hat/assets/js/apps/Iaso/domains/orgUnits/components/OrgUnitsMap.tsx
@@ -190,6 +190,7 @@ export const OrgUnitsMap: FunctionComponent<Props> = ({
             </Grid>
         );
     }
+    console.log('render');
     return (
         <Grid container spacing={0}>
             <InnerDrawer
